+++ src/config/backend/EdFi.DmsConfigurationService.Backend.Postgresql/Repositories/VendorRepository.cs
@@ -6,8 +6,8 @@
 using Dapper;
 using EdFi.DmsConfigurationService.Backend.Repositories;
 using EdFi.DmsConfigurationService.DataModel;
-using EdFi.DmsConfigurationService.DataModel.Application;
-using EdFi.DmsConfigurationService.DataModel.Vendor;
+using EdFi.DmsConfigurationService.DataModel.Model.Application;
+using EdFi.DmsConfigurationService.DataModel.Model.Vendor;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Npgsql;
