+++ src/config/backend/EdFi.DmsConfigurationService.Backend/Repositories/IApplicationRepository.cs
@@ -4,7 +4,7 @@
 // See the LICENSE and NOTICES files in the project root for more information.
 
 using EdFi.DmsConfigurationService.DataModel;
-using EdFi.DmsConfigurationService.DataModel.Application;
+using EdFi.DmsConfigurationService.DataModel.Model.Application;
 
 namespace EdFi.DmsConfigurationService.Backend.Repositories;
