+++ src/config/datamodel/EdFi.DmsConfigurationService.DataModel/Model/Action/ActionResponse.cs
@@ -5,7 +5,7 @@
 
 using System.Text.Json.Serialization;
 
-namespace EdFi.DmsConfigurationService.DataModel.Action;
+namespace EdFi.DmsConfigurationService.DataModel.Model.Action;
 
 public class AdminAction
 {
