+++ src/config/datamodel/EdFi.DmsConfigurationService.DataModel/Model/Application/ApplicationResponse.cs
@@ -3,7 +3,7 @@
 // The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
 // See the LICENSE and NOTICES files in the project root for more information.
 
-namespace EdFi.DmsConfigurationService.DataModel.Application;
+namespace EdFi.DmsConfigurationService.DataModel.Model.Application;
 
 public class ApplicationResponse
 {
