+++ src/config/frontend/EdFi.DmsConfigurationService.Frontend.AspNetCore.Tests.Unit/Modules/IdentityModuleTests.cs
@@ -8,8 +8,8 @@ using System.Net.Http.Json;
 using System.Text.Json.Nodes;
 using EdFi.DmsConfigurationService.Backend;
 using EdFi.DmsConfigurationService.Backend.Repositories;
-using EdFi.DmsConfigurationService.DataModel.Register;
-using EdFi.DmsConfigurationService.DataModel.Token;
+using EdFi.DmsConfigurationService.DataModel.Model.Register;
+using EdFi.DmsConfigurationService.DataModel.Model.Token;
 using EdFi.DmsConfigurationService.Frontend.AspNetCore.Configuration;
 using FakeItEasy;
 using FluentAssertions;
