+++ src/config/frontend/EdFi.DmsConfigurationService.Frontend.AspNetCore/Modules/ActionsModule.cs
@@ -3,7 +3,7 @@
 // The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
 // See the LICENSE and NOTICES files in the project root for more information.
 
-using EdFi.DmsConfigurationService.DataModel.Action;
+using EdFi.DmsConfigurationService.DataModel.Model.Action;
 using EdFi.DmsConfigurationService.Frontend.AspNetCore.Infrastructure;
 
 namespace EdFi.DmsConfigurationService.Frontend.AspNetCore.Modules;
